Move Unity Catalog catalog segments to the Databricks dialect
