+++ src/sqlfluff/dialects/dialect_sparksql.py
@@ -805,38 +805,8 @@ class DatatypeSegment(PrimitiveTypeSegment):
     )
 
 
-# Object Reference Segments
-class CatalogReferenceSegment(ansi.ObjectReferenceSegment):
-    """A reference to a catalog.
-
-    https://docs.databricks.com/data-governance/unity-catalog/create-catalogs.html
-    """
-
-    type = "catalog_reference"
-
-
 # Data Definition Statements
 # http://spark.apache.org/docs/latest/sql-ref-syntax-ddl.html
-class AlterCatalogStatementSegment(BaseSegment):
-    """An `ALTER CATALOG` statement.
-
-    https://docs.databricks.com/sql/language-manual/sql-ref-syntax-ddl-alter-catalog.html
-    """
-
-    type = "alter_catalog_statement"
-    match_grammar = Sequence(
-        "ALTER",
-        "CATALOG",
-        Ref("CatalogReferenceSegment"),
-        Ref.keyword("SET", optional=True),
-        Sequence(
-            "OWNER",
-            "TO",
-            Ref("SingleIdentifierGrammar"),
-        ),
-    )
-
-
 class AlterDatabaseStatementSegment(BaseSegment):
     """An `ALTER DATABASE/SCHEMA` statement.
 
@@ -1062,22 +1032,6 @@ class AlterViewStatementSegment(BaseSegment):
     )
 
 
-class CreateCatalogStatementSegment(BaseSegment):
-    """A `CREATE CATALOG` statement.
-
-    https://docs.databricks.com/sql/language-manual/sql-ref-syntax-ddl-create-catalog.html
-    """
-
-    type = "create_catalog_statement"
-    match_grammar = Sequence(
-        "CREATE",
-        "CATALOG",
-        Ref("IfNotExistsGrammar", optional=True),
-        Ref("CatalogReferenceSegment"),
-        Ref("CommentGrammar", optional=True),
-    )
-
-
 class CreateDatabaseStatementSegment(ansi.CreateDatabaseStatementSegment):
     """A `CREATE DATABASE` statement.
 
@@ -1265,22 +1219,6 @@ class RemoveWidgetStatementSegment(BaseSegment):
     )
 
 
-class DropCatalogStatementSegment(BaseSegment):
-    """A `DROP CATALOG` statement.
-
-    https://docs.databricks.com/sql/language-manual/sql-ref-syntax-ddl-drop-catalog.html
-    """
-
-    type = "drop_catalog_statement"
-    match_grammar: Matchable = Sequence(
-        "DROP",
-        "CATALOG",
-        Ref("IfExistsGrammar", optional=True),
-        Ref("CatalogReferenceSegment"),
-        Ref("DropBehaviorGrammar", optional=True),
-    )
-
-
 class DropDatabaseStatementSegment(ansi.DropDatabaseStatementSegment):
     """A `DROP DATABASE` statement.
 
@@ -1340,20 +1278,6 @@ class TruncateStatementSegment(ansi.TruncateStatementSegment):
     )
 
 
-class UseCatalogStatementSegment(BaseSegment):
-    """A `USE CATALOG` statement.
-
-    https://docs.databricks.com/sql/language-manual/sql-ref-syntax-ddl-use-catalog.html
-    """
-
-    type = "use_catalog_statement"
-    match_grammar = Sequence(
-        "USE",
-        "CATALOG",
-        Ref("CatalogReferenceSegment"),
-    )
-
-
 class UseDatabaseStatementSegment(BaseSegment):
     """A `USE DATABASE` statement.
 
@@ -1364,7 +1288,6 @@ class UseDatabaseStatementSegment(BaseSegment):
 
     match_grammar = Sequence(
         "USE",
-        OneOf("DATABASE", "SCHEMA", optional=True),
         Ref("DatabaseReferenceSegment"),
     )
 
@@ -2573,11 +2496,6 @@ class StatementSegment(ansi.StatementSegment):
             # Databricks - widgets
             Ref("CreateWidgetStatementSegment"),
             Ref("RemoveWidgetStatementSegment"),
-            # Databricks - Unity Catalog
-            Ref("AlterCatalogStatementSegment"),
-            Ref("CreateCatalogStatementSegment"),
-            Ref("DropCatalogStatementSegment"),
-            Ref("UseCatalogStatementSegment"),
         ],
         remove=[
             Ref("TransactionStatementSegment"),
