+++ test/fixtures/dialects/databricks/alter_catalog.sql
