+++ test/fixtures/dialects/databricks/drop_catalog.sql
