+++ test/fixtures/dialects/databricks/use_catalog.sql
