+++ test/fixtures/dialects/databricks/use_database.sql
@@ -1,0 +1,12 @@
+USE database_name;
+
+-- Use the 'userdb'
+USE userdb;
+
+-- Use the 'userdb1'
+USE userdb1;
+
+-- Keywords SCHEMA and DATABASE are interchangeable.
+USE DATABASE database_name;
+
+USE SCHEMA database_name;
