+++ test/fixtures/dialects/sparksql/use_database.sql
@@ -5,8 +5,3 @@ USE userdb;
 
 -- Use the 'userdb1'
 USE userdb1;
-
--- Keywords SCHEMAS and DATABASES are interchangeable.
-USE DATABASE database_name;
-
-USE SCHEMA database_name;
